chore (refs T22614): Move initVue to client/js.
