+++ client/js/bundles/core/genericBundleValidate.js
@@ -16,7 +16,7 @@
  */
 
 import dpValidate from '@demos-europe/demosplan-utils/lib/validation/dpValidate'
-import { initialize } from '@DemosPlanCoreBundle/InitVue'
+import { initialize } from '@DpJs/InitVue'
 
 initialize().then(() => {
   dpValidate()
